requirements.txt
pdf2image==1.15.1
Pillow==8.1.0
pytesseract==0.3.7